requirement.txt
fastapi
uvicorn
jinja2
pdfplumber
python-multipart
google-genai

